requirements.txt
# Core dependencies
pandas>=1.0.0
numpy>=1.18.0
python-dotenv>=0.15.0
pydantic==2.4.2

# Data handling
openpyxl==3.1.2
pyarrow==14.0.1
xlrd==2.0.1
fastparquet==2023.8.0

# AI and ML
openai==1.3.7
scikit-learn==1.3.0
joblib==1.3.2

# Visualization
matplotlib>=3.0.0
seaborn>=0.11.0

# Web API
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
starlette==0.27.0

# Snowflake
snowflake-connector-python>=2.7.0

# Reporting
jinja2>=2.11.0
weasyprint>=53.0
pdfkit==1.0.0

# Scheduling
apache-airflow==2.7.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
requests==2.31.0
PyYAML==6.0.1
pillow>=8.0.0
